requirements/dev.txt
# This file pulls in everything a developer needs. If it's a basic package
# needed to run the site, it belongs in requirements/prod.txt. If it's a
# package for developers (testing, docs, etc.), it goes in this file.
-r prod.txt

# Deployment
Fabric

# Documentation
# Sphinx

# Testing
nose
mock
selenium
django-nose
-e git://github.com/jbalogh/test-utils.git#egg=test-utils
coverage
django-coverage
factory_boy
watchdog

# Shell
# ipython

